I lied, compile cant be fixed until IRC-stuff is further developed - however, I did some cleanup following TC standards, if you dont mind :P

+++ src/server/game/Entities/Unit/Unit.cpp
@@ -3649,7 +3649,7 @@ void Unit::RemoveAurasDueToSpellByDispel(uint32 spellId, uint64 casterGUID, Unit
                     {
                         // final heal
                         int32 healAmount = aurEff->GetAmount();
-                        uint8 stack = chargesRemoved;
+                        int32 stack = chargesRemoved;
                         CastCustomSpell(this, 33778, &healAmount, &stack, NULL, true, NULL, NULL, aura->GetCasterGUID());
 
                         // mana return to caster
